requirements.txt
streamlit
pandas
psycopg2-binary
google-generativeai
supabase
st-supabase-connection
streamlit-authenticator
bcrypt
PyJWT
